Abstract the test case
