+++ composer.json
@@ -25,6 +25,11 @@
 			"Automattic\\WooCommerce\\Blocks\\": "src/"
 		}
 	},
+	"autoload-dev": {
+		"psr-4": {
+			"Automattic\\WooCommerce\\Blocks\\Tests\\": "tests/php"
+		}
+	},
 	"scripts": {
 		"phpcs": [
 			"phpcs --extensions=php -s -p"
