+++ tests/php/StoreApi/Routes/ProductAttributeTerms.php
@@ -3,26 +3,16 @@
  * Controller Tests.
  */
 
-namespace Automattic\WooCommerce\Blocks\Tests\StoreApi\Controllers;
+namespace Automattic\WooCommerce\Blocks\Tests\StoreApi\Routes;
 
+use Automattic\WooCommerce\Blocks\Tests\StoreApi\Routes\ControllerTestCase;
 use Automattic\WooCommerce\Blocks\Tests\Helpers\FixtureData;
 use Automattic\WooCommerce\Blocks\Tests\Helpers\ValidateSchema;
-use Automattic\WooCommerce\Blocks\Domain\Services\ExtendRestApi;
-use Automattic\WooCommerce\Blocks\Package;
-use Automattic\WooCommerce\Blocks\Domain\Package as DomainPackage;
-use Automattic\WooCommerce\Blocks\StoreApi\Formatters;
-use Automattic\WooCommerce\Blocks\StoreApi\Formatters\MoneyFormatter;
-use Automattic\WooCommerce\Blocks\StoreApi\Formatters\HtmlFormatter;
-use Automattic\WooCommerce\Blocks\StoreApi\Formatters\CurrencyFormatter;
-use Automattic\WooCommerce\Blocks\Registry\Container;
-use Automattic\WooCommerce\Blocks\Domain\Services\FeatureGating;
 
 /**
  * Product Attributes Controller Tests.
  */
-class ProductAttributeTerms extends \WP_Test_REST_TestCase {
-
-	private $mock_extend;
+class ProductAttributeTerms extends ControllerTestCase {
 
 	/**
 	 * Setup test products data. Called before every test.
@@ -30,34 +20,12 @@ class ProductAttributeTerms extends \WP_Test_REST_TestCase {
 	public function setUp() {
 		parent::setUp();
 
-		/** @var \WP_REST_Server $wp_rest_server */
-		global $wp_rest_server;
-		$wp_rest_server = new \Spy_REST_Server;
-		do_action( 'rest_api_init', $wp_rest_server );
-
-		wp_set_current_user( 0 );
-		$formatters = new Formatters();
-		$formatters->register( 'money', MoneyFormatter::class );
-		$formatters->register( 'html', HtmlFormatter::class );
-		$formatters->register( 'currency', CurrencyFormatter::class );
-		$this->mock_extend = new ExtendRestApi( new DomainPackage( '', '', new FeatureGating( 2 ) ), $formatters );
-
 		$fixtures = new FixtureData();
 
 		$this->attributes = [
 			$fixtures->get_product_attribute( 'color', [ 'red', 'green', 'blue' ] ),
 			$fixtures->get_product_attribute( 'size', [ 'small', 'medium', 'large' ] )
 		];
-	}
-
-	/**
-	 * Tear down the above.
-	 */
-	public function tearDown() {
-		parent::tearDown();
-		/** @var \WP_REST_Server $wp_rest_server */
-		global $wp_rest_server;
-		$wp_rest_server = null;
 	}
 
 	/**
